CORE-8643 Move ArgumentProperties from dialog package to model
