+++ de-lib/src/main/java/org/iplantc/de/apps/integration/client/dialogs/CommandLineOrderingPanel.java
@@ -1,5 +1,6 @@
 package org.iplantc.de.apps.integration.client.dialogs;
 
+import org.iplantc.de.apps.integration.client.model.ArgumentProperties;
 import org.iplantc.de.apps.integration.client.view.AppsEditorView;
 import org.iplantc.de.client.models.apps.integration.Argument;
 import org.iplantc.de.resources.client.uiapps.integration.AppIntegrationMessages;
