+++ de-lib/src/main/java/org/iplantc/de/apps/integration/client/model/ArgumentProperties.java
@@ -1,4 +1,4 @@
-package org.iplantc.de.apps.integration.client.dialogs;
+package org.iplantc.de.apps.integration.client.model;
 
 import org.iplantc.de.client.models.apps.integration.Argument;
